requirements.txt
python-dotenv==1.0.1
apscheduler==3.10.4
pytz==2024.1
python-dateutil==2.9.0.post0
requests==2.32.3
pyyaml==6.0.2
feedparser==6.0.11
markdown==3.5.1